requirements.txt
streamlit==1.32.0
pandas==2.2.1
python-dotenv==1.0.1
langchain==0.1.12
langchain-openai==0.0.8
langgraph==0.0.27
openai==1.14.0
plotly==5.19.0 